puppetmaster: improve error handling for puppet-facts-upload
Bug: T311742

+++ modules/puppetmaster/files/puppet-facts-upload.py
@@ -19,10 +19,12 @@ from argparse import ArgumentParser
 def get_args():
     """Parse arguments"""
     parser = ArgumentParser(description=__doc__)
-    parser.add_argument('-u', '--uri', default='https://puppet-compiler.wmflabs.org/upload')
-    parser.add_argument('-s', '--skip', action='store_true')
-    parser.add_argument('-p', '--proxy')
-    parser.add_argument('-v', '--verbose', action='count')
+    parser.add_argument(
+        "-u", "--uri", default="https://puppet-compiler.wmflabs.org/upload"
+    )
+    parser.add_argument("-s", "--skip", action="store_true")
+    parser.add_argument("-p", "--proxy")
+    parser.add_argument("-v", "--verbose", action="count")
     return parser.parse_args()
 
 
@@ -37,7 +39,7 @@ def get_log_level(args_level):
 
 
 def get_signature(key, payload):
-    logging.debug('Generating signiture')
+    logging.debug("Generating signiture")
     private_key = serialization.load_pem_private_key(
         key, password=None, backend=default_backend()
     )
@@ -51,35 +53,38 @@ def get_signature(key, payload):
             hashes.SHA256(),
         )
     )
-    logging.debug('Generated signiture: %s', signature)
+    logging.debug("Generated signiture: %s", signature)
     return signature
 
 
 def get_realm():
-    wmcs_projects_file = Path('/etc/wmcs-project')
+    wmcs_projects_file = Path("/etc/wmcs-project")
     if wmcs_projects_file.is_file():
         return wmcs_projects_file.read_text().strip()
-    return 'production'
+    return "production"
 
 
 def generate_facts_file(facts_file, skip):
     if not skip:
         if facts_file.is_file():
-            logging.debug('deleting old file: %s', facts_file)
+            logging.debug("deleting old file: %s", facts_file)
             facts_file.unlink()
-        logging.debug('Generating facts export: %s', facts_file)
-        subprocess.run('/usr/local/bin/puppet-facts-export', check=True)
+        logging.debug("Generating facts export: %s", facts_file)
+        subprocess.run("/usr/local/bin/puppet-facts-export", check=True)
     if not facts_file.is_file():
-        logging.error('Error creating facts export')
+        logging.error("Error creating facts export")
         raise SystemExit(1)
 
 
 def get_key_file():
-    return Path(subprocess.run(
-        'facter -p puppet_config.hostprivkey'.split(),
-        check=True,
-        capture_output=True,
-        ).stdout.decode().strip()
+    return Path(
+        subprocess.run(
+            "facter -p puppet_config.hostprivkey".split(),
+            check=True,
+            capture_output=True,
+        )
+        .stdout.decode()
+        .strip()
     )
 
 
@@ -88,26 +93,35 @@ def main():
     args = get_args()
     logging.basicConfig(level=get_log_level(args.verbose))
 
-    facts_file = Path('/tmp/puppet-facts-export.tar.xz')
+    facts_file = Path("/tmp/puppet-facts-export.tar.xz")
     generate_facts_file(facts_file, args.skip)
 
     key_file = get_key_file()
     signature = get_signature(key_file.read_bytes(), facts_file.read_bytes())
 
-    data = {'realm': get_realm(), 'hostname': gethostname(), 'signature': signature}
-    files = {'file': facts_file.read_bytes()}
-    proxies = {'http': args.proxy, 'https': args.proxy} if args.proxy else None
+    data = {"realm": get_realm(), "hostname": gethostname(), "signature": signature}
+    files = {"file": facts_file.read_bytes()}
+    proxies = {"http": args.proxy, "https": args.proxy} if args.proxy else None
 
-    logging.debug('Posting data to: %s', args.uri)
+    logging.debug("Posting data to: %s", args.uri)
     resp = post(args.uri, data=data, files=files, proxies=proxies)
-    if not resp.json().get('result'):
-        logging.error('Error submitting facts file')
+    if not resp.ok:
+        if resp.status_code == 403:
+            logging.error(
+                "request denied, ensure you have added the puppetmaster certificate "
+                "https://wikitech.wikimedia.org/wiki/Help:Puppet-compiler#Manually_update_cloud"
+            )
+        else:
+            logging.error("%d: Unable to submit request: %s", resp.status_code, resp.text)
+        return 1
+    if not resp.json().get("result"):
+        logging.error("Error submitting facts file")
         return 1
 
-    logging.debug('Success!')
+    logging.debug("Success!")
     facts_file.unlink()
     return 0
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     raise SystemExit(main())
